Resolve "Fix Activity class"

Closes #43

See merge request cse1105/2021-2022/team-repositories/oopp-group-26/repository-template!46

+++ client/src/main/java/client/scenes/MainCtrl.java
@@ -272,7 +272,7 @@ public class MainCtrl implements MainCtrlRequirements {
      * @param baseScore the score in range [0; 100]
      */
     @Override
-    public void addPoints(int baseScore) {
+    public void addPoints(long baseScore) {
         if (baseScore != 0) {
             double progress = ((double) (timeUtils.now() - questionStartTime)) / (questionEndTime - questionStartTime);
             pointsGained = (int) (50.0 + 0.5 * (1.0 - progress) * (double) baseScore);
