+++ client/src/main/java/client/scenes/controllerrequirements/MainCtrlRequirements.java
@@ -25,7 +25,7 @@ public interface MainCtrlRequirements {
      *
      * @param baseScore the score (0 - 100)
      */
-    void addPoints(int baseScore);
+    void addPoints(long baseScore);
 
     void showGlobalLeaderboardFrame();
