+++ server/src/main/java/server/api/ActivityController.java
@@ -185,8 +185,8 @@ public class ActivityController {
         int correctAnswer = 0;
 
         if (numberOfActivities == 1) {
-            int correctNumber = activities.get(0).consumptionInWh;
-            int wrongNumber = correctNumber * 110 / 100;
+            long correctNumber = activities.get(0).consumptionInWh;
+            long wrongNumber = correctNumber * 110 / 100;
             if (wrongNumber % 2 == 0) {
                 question = activities.get(0).title + " consumes " + wrongNumber + "per hour.";
                 correctAnswer = 1;
